requirements.txt
flask
flask-sqlalchemy
ipython
psycopg2-binary
sqlalchemy
flask-bcrypt
email_validator
flask_wtforms